pyproject.toml
[tool.poetry]
name = "transcrib"
version = "0.1.0"
description = ""
authors = ["VeterAN <ver-an@yandex.ru>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
openai-whisper = "^20231117"
python-dotenv = "^1.0.1"
loguru = "^0.7.2"
transformers = "^4.37.2"
pydub = "^0.25.1"
pillow = "^10.2.0"
sentencepiece = "^0.1.99"
sacremoses = "^0.1.1"
ffmpeg = "^1.4"
ffmpeg-python = "^0.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79
target-version = ['py310']
